RFC remove file util.py and add Benchmark
