+++ benchopt/base.py
@@ -1,196 +1,15 @@
 import tempfile
 from abc import ABC, abstractmethod
 
-from .util import product_param
-from .utils.class_property import classproperty
 from .utils.dynamic_modules import get_file_hash
-from .utils.shell_cmd import install_in_conda_env
-from .utils.shell_cmd import _run_shell_in_conda_env
 from .utils.dynamic_modules import _reconstruct_class
-from .utils.shell_cmd import shell_install_in_conda_env
 
-from .config import RAISE_INSTALL_ERROR
+from .utils.dependencies_mixing import DependenciesMixin
+from .utils.parametrized_name_mixing import ParametrizedNameMixin
 
 
 # Possible stop strategies
 STOP_STRATEGIES = ['iteration', 'tolerance']
-
-
-class ParametrizedNameMixin():
-    """Mixing for parametric classes representation and naming.
-    """
-    parameters = {}
-
-    def __init__(self, **parameters):
-        """Default init set parameters base on the cls.parameters
-        """
-        parameters_ = next(product_param(self.parameters))
-        parameters_.update(parameters)
-        for k, v in parameters_.items():
-            if not hasattr(self, k):
-                setattr(self, k, v)
-
-    def save_parameters(self, **parameters):
-        self.parameters = parameters
-        if not hasattr(self, 'parameter_template'):
-            self.parameter_template = ",".join(
-                [f"{k}={v}" for k, v in parameters.items()])
-        for k, v in parameters.items():
-            if not hasattr(self, k):
-                setattr(self, k, v)
-
-    @classmethod
-    def get_instance(cls, **parameters):
-        """Helper function to instantiate an object and save its parameters.
-
-        Saving the parameters allow for cheap hashing and to compute parametric
-        names for the objects.
-        """
-        obj = cls(**parameters)
-        obj.save_parameters(**parameters)
-        return obj
-
-    @property
-    @abstractmethod
-    def name(self):
-        """Each object should expose its name for plotting purposes."""
-        ...
-
-    def __repr__(self):
-        """Compute the parametrized name of the instance."""
-        out = f"{self.name}"
-        if len(self.parameters) > 0:
-            out += f"[{self.parameter_template}]".format(**self.parameters)
-        return out
-
-    @classmethod
-    def _get_parametrized_name(cls, **parameters):
-        """Compute the parametrized name for a given set of parameters."""
-        return str(cls.get_instance(**parameters))
-
-    @classmethod
-    def _reload_class(cls, pickled_module_hash=None):
-
-        return _reconstruct_class(
-            cls._module_filename, cls._base_class_name,
-            pickled_module_hash=pickled_module_hash
-        )
-
-
-class DependenciesMixin:
-    # Information on how to install the class. The value of install_cmd should
-    # be in {None, 'conda', 'shell'}. The API reads:
-    #
-    # - 'conda': The class should have an attribute `requirements`.
-    #          BenchOpt will conda install `$requirements`, except for entries
-    #          starting with `pip:` which will be installed with `pip` in the
-    #          conda env.
-    #
-    # - 'shell': The solver should have attribute `install_script`. BenchOpt
-    #           will run `install_script` in a shell and provide the conda
-    #           env directory as an argument. The command should then be
-    #           installed in the `bin` folder of the env and can be imported
-    #           with import_shell_cmd in the safe_import_context.
-    install_cmd = None
-
-    @classproperty
-    def benchmark(cls):
-        return cls.__module__.split('.')[1]
-
-    @classproperty
-    def name(cls):
-        return cls.__module__.split('.')[-1]
-
-    @classmethod
-    def is_installed(cls, env_name=None, raise_on_not_installed=None):
-        """Check if the module caught a failed import to assert install.
-
-        Parameters
-        ----------
-        env_name: str or None
-            Name of the conda env where the install should be checked. If
-            None, check the install in the current environment.
-        raise_on_not_installed: boolean or None
-            If set to True, raise an error if the requirements are not
-            installed. This is mainly for testing purposes.
-
-        Returns
-        -------
-        is_installed: bool
-            returns True if no import failure has been detected.
-        """
-        if env_name is None:
-            if (cls._import_ctx is not None
-                    and cls._import_ctx.failed_import):
-                if raise_on_not_installed:
-                    exc_type, value, tb = cls._import_ctx.import_error
-                    raise exc_type(value).with_traceback(tb)
-                return False
-            else:
-                return True
-        else:
-            return _run_shell_in_conda_env(
-                f"benchopt check-install {cls._module_filename} "
-                f"{cls._base_class_name}",
-                env_name=env_name, raise_on_error=raise_on_not_installed
-            ) == 0
-
-    @classmethod
-    def install(cls, env_name=None, force=False):
-        """Install the class in the given conda env.
-
-        Parameters
-        ----------
-        env_name: str or None
-            Name of the conda env where the class should be installed. If
-            None, tries to install it in the current environment.
-        force : boolean (default: False)
-            If set to True, forces reinstallation when using conda.
-
-        Returns
-        -------
-        is_installed: bool
-            True if the class is correctly installed in the environment.
-        """
-        is_installed = cls.is_installed(env_name=env_name)
-
-        if force or not is_installed:
-            print(f"Installing {cls.name} in {env_name}:...",
-                  end='', flush=True)
-            try:
-                cls._pre_install_hook(env_name=env_name)
-                if cls.install_cmd == 'conda':
-                    install_in_conda_env(*cls.requirements, env_name=env_name,
-                                         force=force)
-                elif cls.install_cmd == 'shell':
-                    install_file = (
-                        cls._module_filename.parents[1] / 'install_scripts' /
-                        cls.install_script
-                    )
-                    shell_install_in_conda_env(install_file, env_name=env_name)
-                cls._post_install_hook(env_name=env_name)
-
-            except Exception as exception:
-                if RAISE_INSTALL_ERROR:
-                    raise exception
-
-            is_installed = cls.is_installed(env_name=env_name)
-            if is_installed:
-                print(" done")
-            else:
-                print(" failed")
-
-        return is_installed
-
-    @classmethod
-    def _pre_install_hook(cls, env_name=None):
-        """Hook called before installing dependencies with conda or pip."""
-        pass
-
-    @classmethod
-    def _post_install_hook(cls, env_name=None):
-        """Hook called after installing dependencies with conda or pip."""
-        pass
 
 
 class BaseSolver(ParametrizedNameMixin, DependenciesMixin, ABC):
@@ -267,6 +86,15 @@ class BaseSolver(ParametrizedNameMixin, DependenciesMixin, ABC):
         """
         ...
 
+    @classmethod
+    def supports_dataset(cls, dataset):
+        # Check that the solver is compatible with the given dataset
+        if (getattr(dataset, 'is_sparse', False)
+                and not getattr(cls, 'support_sparse', True)):
+            return False
+
+        return True
+
     # TODO: use this to allow parallel computation of the benchmark.
     @staticmethod
     def _reconstruct(module_filename, parameters, objective,
@@ -301,7 +129,7 @@ class CommandLineSolver(BaseSolver, ABC):
         super().__init__(**parameters)
 
 
-class BaseDataset(ParametrizedNameMixin, DependenciesMixin):
+class BaseDataset(ParametrizedNameMixin, DependenciesMixin, ABC):
     """Base class to define a dataset in a benchmark.
 
     Datasets that derive from this class should implement one method:
