+++ benchopt/datasets/__init__.py
@@ -1,0 +1,4 @@
+
+from .simulated import make_correlated_data
+
+__all__ = ['make_correlated_data']
