+++ benchopt/tests/test_benchmarks/dummy_benchmark/datasets/boston.py
@@ -1,6 +1,5 @@
-from benchopt.base import BaseDataset
-
-from benchopt.util import safe_import_context
+from benchopt import BaseDataset
+from benchopt import safe_import_context
 
 
 with safe_import_context() as import_ctx:
