+++ benchopt/tests/test_benchmarks/dummy_benchmark/solvers/celer.py
@@ -1,7 +1,7 @@
 import warnings
 
-from benchopt.base import BaseSolver
-from benchopt.util import safe_import_context
+from benchopt import BaseSolver
+from benchopt import safe_import_context
 
 
 with safe_import_context() as import_ctx:
