+++ benchopt/utils/__init__.py
@@ -1,0 +1,6 @@
+
+from .safe_import import safe_import_context
+from .parametrized_name_mixing import product_param
+
+
+__all__ = ["safe_import_context", "product_param"]
