+++ benchopt/utils/safe_import.py
@@ -1,0 +1,36 @@
+import warnings
+
+
+from ..config import RAISE_INSTALL_ERROR
+
+
+class safe_import_context:
+    """Do not fail on ImportError and catch import warnings"""
+
+    def __init__(self):
+        self.failed_import = False
+        self.record = warnings.catch_warnings(record=True)
+
+    def __enter__(self):
+        # Catch the import warning except if install errors are raised.
+        if not RAISE_INSTALL_ERROR:
+            self.record.__enter__()
+        return self
+
+    def __exit__(self, exc_type, exc_value, tb):
+
+        silence_error = False
+
+        # prevent import error from propagating and tag
+        if exc_type is not None and issubclass(exc_type, ImportError):
+            self.failed_import = True
+            self.import_error = exc_type, exc_value, tb
+
+            # Prevent the error propagation
+            silence_error = True
+
+        if not RAISE_INSTALL_ERROR:
+            self.record.__exit__(exc_type, exc_value, tb)
+
+        # Returning True in __exit__ prevent error propagation.
+        return silence_error
